+++ CHANGES.rst
@@ -1,3 +1,8 @@
+UNRELEASED
+----------
+
+- Add support for connecting to the Tor control port over a unix domain socket.
+
 0.1.4
 -----